Bug Fix : When Package Weight is greater than max_carriable_weight
The application goes into an infinite loop if the weight of the package is greater than max weight that the delivery agent can carry. As part of the changes if this happens it is marked as unavailable during delivery time computation though it would still compute the discount and the delivery cost.
Also a test is added in the C++ solution to capture the earlier failing case.